ssh_auth_podman_push
Module generic_functions_mod

Generic functions that does not usually change.

Don’t change this code, so it can be updated regularly with cargo auto update_automation_tasks_rs If you want to customize it, copy the code into main.rs and modify it there.

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color

Functions

panic_set_hook

The original Rust report of the panic is ugly for the end user

tracing_init

Initialize tracing to file logs/automation_tasks_rs.log